requirements.txt
numpy
pandas
akshare
yfinance
torch
torchvision
pillow
faiss-cpu
scikit-learn
scipy
langchain
langchain-google-genai
langchain-community
pydantic
streamlit
plotly
matplotlib
python-dotenv
pyyaml
tqdm
mplfinance